// src/types/models/add-partial-transcript.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI Template
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import { RealtimeRecognitionMetadata } from './realtime-recognition-metadata';
// May contain unused imports in some cases
// @ts-ignore
import { RealtimeRecognitionResult } from './realtime-recognition-result';

/**
 *
 * @export
 * @interface AddPartialTranscript
 */
export interface AddPartialTranscript {
  /**
   * Speechmatics JSON output format version number.
   * @type {string}
   * @memberof AddPartialTranscript
   */
  format?: string;
  /**
   *
   * @type {string}
   * @memberof AddPartialTranscript
   */
  message: AddPartialTranscriptMessageEnum;
  /**
   *
   * @type {RealtimeRecognitionMetadata}
   * @memberof AddPartialTranscript
   */
  metadata: RealtimeRecognitionMetadata;
  /**
   *
   * @type {Array<RealtimeRecognitionResult>}
   * @memberof AddPartialTranscript
   */
  results: Array<RealtimeRecognitionResult>;
}

export const AddPartialTranscriptMessageEnum = {
  AddPartialTranscript: 'AddPartialTranscript',
} as const;

export type AddPartialTranscriptMessageEnum =
  typeof AddPartialTranscriptMessageEnum[keyof typeof AddPartialTranscriptMessageEnum];